sgpp coding days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2719 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/combigrid/converter/CombiSGppConverter.cpp
@@ -46,7 +46,7 @@ void combigrid::CombiSGppConverter::createSGpp( sg::base::GridStorage* storage ,
 				storage->insert( (*hgi) );
 				//COMBIGRID_OUT_LEVEL3(4 , "createSGpp  add point nrp:" << nrp << " , sgppIndex:" << sgppIndex);
 			}
-			sgppIndex = (*storage)[hgi];
+			sgppIndex = static_cast<int>((*storage)[hgi]);
 			fg->getSGppIndex()[nrp] = sgppIndex;
 		}
 	}
@@ -211,7 +211,7 @@ void combigrid::CombiSGppConverter::SGppToFullGrid( sg::base::GridStorage* stora
 			// rehash for this index
 			hgi->rehash();
 			//todo: we do not test if this is present in the hashmap
-			sgppIndex = (*storage)[hgi];
+			sgppIndex = static_cast<int>((*storage)[hgi]);
 			fg->getElementVector()[nrp] = (*alpha)[sgppIndex];
 			fg->getSGppIndex()[nrp] = sgppIndex;
 		}
